Replace the old JSX builder with the one that has been used experimentally
