--- packages/babel-helper-builder-react-jsx-experimental/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "@babel/helper-builder-react-jsx-experimental",
-  "version": "7.12.4",
-  "description": "Helper function to build react jsx",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/babel/babel.git",
-    "directory": "packages/babel-helper-builder-react-jsx-experimental"
-  },
-  "license": "MIT",
-  "publishConfig": {
-    "access": "public"
-  },
-  "main": "lib/index.js",
-  "dependencies": {
-    "@babel/helper-annotate-as-pure": "workspace:^7.10.4",
-    "@babel/helper-module-imports": "workspace:^7.12.1",
-    "@babel/types": "workspace:^7.12.1"
-  }
-}
